+++ src/base/components/texts/Superscript.tsx
@@ -8,10 +8,11 @@
 
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
+import type { ReactNode } from 'react';
 
 interface SuperscriptProps {
     superscript: string;
-    text: string;
+    text: ReactNode;
 }
 
 /**