+++ src/features/reader/overlay/navigation/desktop/quick-settings/components/ReaderNavBarDesktopReadingDirection.tsx
@@ -7,7 +7,6 @@
  */
 
 import { useLingui } from '@lingui/react/macro';
-import { ValueRotationButton } from '@/base/components/buttons/ValueRotationButton.tsx';
 import type { IReaderSettingsWithDefaultFlag, ReadingDirection } from '@/features/reader/Reader.types.ts';
 import {
     READING_DIRECTION_VALUES,
@@ -15,6 +14,7 @@ import {
 } from '@/features/reader/settings/ReaderSettings.constants.tsx';
 
 import type { SelectButtonDefaultableProps } from '@/base/components/buttons/SelectButton.tsx';
+import { SelectButton } from '@/base/components/buttons/SelectButton.tsx';
 
 export const ReaderNavBarDesktopReadingDirection = ({
     readingDirection,
@@ -27,7 +27,7 @@ export const ReaderNavBarDesktopReadingDirection = ({
     const { t } = useLingui();
 
     return (
-        <ValueRotationButton
+        <SelectButton
             {...buttonSelectInputProps}
             tooltip={t`Reading direction`}
             value={readingDirection.isDefault ? undefined : readingDirection.value}
@@ -36,6 +36,7 @@ export const ReaderNavBarDesktopReadingDirection = ({
             setValue={setReadingDirection}
             valueToDisplayData={READING_DIRECTION_VALUE_TO_DISPLAY_DATA}
             defaultIcon={READING_DIRECTION_VALUE_TO_DISPLAY_DATA[readingDirection.value].icon}
+            isCollapsible
         />
     );
 };