+++ CHANGELOG.md
@@ -29,7 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Migration**) Change migration match exclude/include icons
 - (**Migration**) Show the exclude/include button only for an entry with a selected match
 - (**Source/Extension**) Rename language "All" to "Multi"
--
+- (**Reader**) Simplify changing settings in desktop sidebar
 
 ### Fixed